moved to new folder, unused urlHashMap
